requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
certifi==2025.10.5
charset-normalizer==3.4.3
chex==0.1.90
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
docker-pycreds==0.4.0
etils==1.13.0
flax==0.9.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
humanize==4.13.0
hydra-core==1.3.2
idna==3.11
importlib_resources==6.5.2
jax==0.4.33
jaxlib==0.4.33
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==1.26.4
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.2
orbax-checkpoint==0.6.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
protobuf==5.29.5
psutil==7.1.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.41.0
setproctitle==1.3.7
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
tensorstore==0.1.78
toolz==1.0.0
tqdm==4.66.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.17.7
zipp==3.23.0
